Fixed some compiler warnings by changing the definition of an meInt to being an int and fixing other warnings this created.
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/trunk@2050 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ microemacs/src/emain.h
@@ -571,8 +571,8 @@ typedef   signed char  meByte ;
 typedef unsigned char  meUByte ;
 typedef   signed short meShort ;
 typedef unsigned short meUShort ;
-typedef   signed long  meInt ;
-typedef unsigned long  meUInt ;
+typedef   signed int   meInt ;
+typedef unsigned int   meUInt ;
 
 /* Fix any default search path */
 #ifndef _DEFAULT_SEARCH_PATH
